% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ow_data.R
\name{load_ow_data}
\alias{load_ow_data}
\title{Load Overwatch data}
\usage{
load_ow_data(profile_name, hero_table, season_choice, file_path)
}
\arguments{
\item{profile_name}{A platform-agnostic profile name (Blizzard BattleTag, Xbox Gamertag, PSN ID).}

\item{hero_table}{Text string specifying which hero to get data for (see hero_names()), or "profile" to get general profile information.}

\item{season_choice}{An integer specifying which competitive season to get data for.}

\item{file_path}{File path to where data has been saved by scrape_ow_data().}
}
\value{
A [tibble][tibble::tibble-package] of records containing stats for competitive play between saved OWAPI queries.
}
\description{
Load data obtained by OWAPI and overwatchr into R sesson for analysis. The returned data describes all play between each OWAPI query, and automatically calculates statistics per 10 minutes.
}
\examples{
load_ow_data("catmaps", "moira", 19, here::here("player_data"))
}
