% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwatchr-package.R
\docType{package}
\name{overwatchr-package}
\alias{overwatchr}
\alias{overwatchr-package}
\title{overwatchr: R Interface to Query, Store, and Analyze OWAPI Data}
\description{
This package provides a suite of functions for the collection of data about public Overwatch accounts from the unofficial OWAPI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benbellman/overwatchr}
  \item Report bugs at \url{https://github.com/benbellman/overwatchr/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Bellman \email{ben.h.bellman@gmail.com}

}
\keyword{internal}
