% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ow_data.R
\name{scrape_ow_data}
\alias{scrape_ow_data}
\title{Scrape OWAPI}
\usage{
scrape_ow_data(profile_name, platform, file_path)
}
\arguments{
\item{profile_name}{Account name of public Overwatch profile (Blizzard BattleTag, Xbox Gamertag, PSN ID). If using a Blizzard BattleTag, "#" will be replaced with "-" in saved data.}

\item{platform}{Profile's gaming platform. One of c("pc", "xbl", "psn"). Nintendo Switch not yet supported.}

\item{file_path}{File path to save/append all hero-specific data files.}
}
\value{
No value returned in R session.
}
\description{
Obtain and save current competitive season statistics for a public Overwatch profile.
}
\examples{
scrape_ow_data("catmaps", "psn", here::here("player_data"))
}
